Move all legacy components into subdirectory
