+++ client/src/components/legacy/Layout.tsx
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 
+
 export const Row = styled.div`
   display: flex;
   align-items: center;
