+++ client/src/components/legacy/ProgressBar.tsx
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 
+
 export interface ProgressBarProps {
   width: number;
   label?: string;
