+++ client/src/components/legacy/ReadOnlyInput.tsx
@@ -1,6 +1,7 @@
 import styled from "styled-components";
 import { Col } from "./Layout";
 
+
 export const ReadOnlyInput: React.FC<{
   label: string;
   value: any;
