+++ client/src/components/legacy/TopBanner.tsx
